iw612-utils.bb: update condition to install NetworkManager conf file
We are using PREFERRED_CONNECTIVITY_MANAGER variable in meta-variscite-bsp-imx/conf/machine/variscite.inc
to define the preferred network connectivity manager, so use it to
conditionally install 99-iw61x-unmanaged-devices.conf.

This avoid to install 99-iw61x-unmanaged-devices.conf when the PREFERRED_CONNECTIVITY_MANAGER
is not configured as networkmanager.

Note:
In Scarthgap release we switched from networkmanager to systemd-networkd.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>