conf: Correct BBFILE_PRIORITY for variscite-bsp-common
The configuration file was setting BBFILE_PRIORITY_variscite-bsl-common
instead of BBFILE_PRIORITY_variscite-bsp-common. This typo in the meta
layer name caused the meta-variscite-bsp-common layer priority to the
default value of 5 instead of the intended 16, leading to some recipes
not being assigned the desired priority.

This fix ensures that the correct layer priority is applied, resolving
potential conflicts in recipe selection.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>