Fix IW612 OpenThread client/server scripts
Update the scripts for managing variscite-ot-client/server to align with the
mandatory use of an OpenThread dataset when starting a Thread connection.

An OpenThread dataset is a collection of configuration parameters required to
define a Thread network's operational settings. These parameters are essential
for creating, joining, or operating within a Thread network and include:
  - Channel: Specifies the 802.15.4 radio channel (e.g., 26).
  - PAN ID: A unique identifier for the Thread network.
  - Extended PAN ID: An extended identifier for differentiating networks.
  - Network Name: The name of the Thread network.
  - Network Key: A 16-byte key for securing Thread communications.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>