openthread: Replace ot-daemon recipe with NXP's openthread recipe
Initially, Variscite added support for OpenThread using ot-daemon recipe,
which needed to be updated to meet specific requirements.

With the Yocto Scarthgap release, NXP introduced the meta-nxp-connectivity
layer, which includes support for the OpenThread Daemon. To align with
NXP changes and reduce maintenance overhead, the ot-daemon recipe is now
deprecated in favor of NXP's implementation.

This commit:
 - Removes the old ot-daemon recipe.
 - Leverages NXP recipes to integrate ot-daemon-iwxxx-spi.
 - Updates variscite-ot-server and variscite-ot-client scripts to use
   ot-client-iwxxx-spi.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>